requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=0.9.0=pypi_0
astor=0.8.1=pypi_0
atomicwrites=1.3.0=pypi_0
attrs=19.3.0=pypi_0
backcall=0.1.0=pypi_0
bleach=1.5.0=pypi_0
certifi=2020.4.5.1=py36h9f0ad1d_0
cffi=1.14.0=py36ha419a9e_0
cloudpickle=1.3.0=pypi_0
colorama=0.4.3=pypi_0
cycler=0.10.0=pypi_0
decorator=4.4.2=pypi_0
defusedxml=0.6.0=pypi_0
docopt=0.6.2=pypi_0
entrypoints=0.3=pypi_0
future=0.18.2=pypi_0
gast=0.3.3=pypi_0
grpcio=1.11.0=pypi_0
gym=0.17.1=pypi_0
html5lib=0.9999999=pypi_0
importlib-metadata=1.6.0=pypi_0
intel-openmp=2020.0=166
ipykernel=5.2.0=pypi_0
ipython=7.13.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.5.1=pypi_0
jedi=0.16.0=pypi_0
jinja2=2.11.1=pypi_0
jsonschema=3.2.0=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=6.1.2=pypi_0
jupyter-console=6.1.0=pypi_0
jupyter-core=4.6.3=pypi_0
kiwisolver=1.2.0=pypi_0
libblas=3.8.0=15_mkl
libcblas=3.8.0=15_mkl
liblapack=3.8.0=15_mkl
markdown=3.2.1=pypi_0
markupsafe=1.1.1=pypi_0
matplotlib=3.2.1=pypi_0
mistune=0.8.4=pypi_0
mkl=2020.0=166
more-itertools=8.2.0=pypi_0
nbconvert=5.6.1=pypi_0
nbformat=5.0.5=pypi_0
notebook=6.0.3=pypi_0
numpy=1.18.2=pypi_0
packaging=20.3=pypi_0
pandas=1.0.3=pypi_0
pandocfilters=1.4.2=pypi_0
parso=0.6.2=pypi_0
pickleshare=0.7.5=pypi_0
pillow=7.1.1=pypi_0
pip=20.0.2=py_2
pluggy=0.13.1=pypi_0
prometheus-client=0.7.1=pypi_0
prompt-toolkit=3.0.5=pypi_0
protobuf=3.5.2=pypi_0
py=1.8.1=pypi_0
pycparser=2.20=py_0
pyglet=1.5.0=pypi_0
pygments=2.6.1=pypi_0
pyparsing=2.4.7=pypi_0
pyrsistent=0.16.0=pypi_0
pytest=5.4.1=pypi_0
python=3.6.10=he025d50_1009_cpython
python-dateutil=2.8.1=pypi_0
python_abi=3.6=1_cp36m
pytorch=0.4.0=py36_cuda80_cudnn7he774522_1
pytz=2019.3=pypi_0
pywin32=227=pypi_0
pywinpty=0.5.7=pypi_0
pyyaml=5.3.1=pypi_0
pyzmq=19.0.0=pypi_0
qtconsole=4.7.2=pypi_0
qtpy=1.9.0=pypi_0
scipy=1.4.1=pypi_0
send2trash=1.5.0=pypi_0
setuptools=46.1.3=py36h9f0ad1d_0
six=1.14.0=pypi_0
tensorboard=1.7.0=pypi_0
tensorflow=1.7.1=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.8.3=pypi_0
testpath=0.4.4=pypi_0
tornado=6.0.4=pypi_0
traitlets=4.3.3=pypi_0
unityagents=0.4.0=pypi_0
vc=14.1=h869be7e_1
vs2015_runtime=14.16.27012=h30e32a0_1
wcwidth=0.1.9=pypi_0
werkzeug=1.0.1=pypi_0
wheel=0.34.2=py_1
widgetsnbextension=3.5.1=pypi_0
wincertstore=0.2=py36_1003
zipp=3.1.0=pypi_0
